refactor: explorer page tab logic separated as component (#3)
